backend/requirements.txt
newspaper4k[all]
pyyaml
openai
python-dotenv
fastapi
uvicorn[standard]
apscheduler